+++ src/main/java/org/opentripplanner/api/resource/LIsochrone.java
@@ -40,7 +40,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.io.Files;
-import com.vividsolutions.jts.geom.MultiPolygon;
+import org.locationtech.jts.geom.MultiPolygon;
 
 /**
  * Return isochrone geometry as a set of GeoJSON or zipped-shapefile multi-polygons.