+++ src/main/java/org/opentripplanner/gtfs/mapping/TripMapper.java
@@ -42,6 +42,13 @@ class TripMapper {
         lhs.setTripBikesAllowed(rhs.getTripBikesAllowed());
         lhs.setBikesAllowed(rhs.getBikesAllowed());
         lhs.setFareId(rhs.getFareId());
+        lhs.setDrtMaxTravelTime(rhs.getDrtMaxTravelTime());
+        lhs.setDrtAvgTravelTime(rhs.getDrtAvgTravelTime());
+        lhs.setDrtAdvanceBookMin(rhs.getDrtAdvanceBookMin());
+        lhs.setDrtPickupMessage(rhs.getDrtPickupMessage());
+        lhs.setDrtDropOffMessage(rhs.getDrtDropOffMessage());
+        lhs.setContinuousPickupMessage(rhs.getContinuousPickupMessage());
+        lhs.setContinuousDropOffMessage(rhs.getContinuousDropOffMessage());
 
         return lhs;
     }