+++ src/main/java/org/opentripplanner/routing/edgetype/SimpleTransfer.java
@@ -7,7 +7,7 @@ import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.vertextype.TransitStop;
 
-import com.vividsolutions.jts.geom.LineString;
+import org.locationtech.jts.geom.LineString;
 
 import java.util.List;
 import java.util.Locale;
@@ -46,6 +46,10 @@ public class SimpleTransfer extends Edge {
             return null;
         }
         if(distance > s0.getOptions().maxTransferWalkDistance) {
+            return null;
+        }
+        // Don't allow SimpleTransfer right after a call-and-ride or deviated-route dropoff - in that case we need to transfer at the same stop
+        if (s0.isLastBoardAlightDeviated()) {
             return null;
         }
         // Only transfer right after riding a vehicle.