+++ docs/Changelog.md
@@ -7,6 +7,8 @@
 - Update onebusaway-gtfs to latest version from OBA project (#2636)
 - Remove the coupling to OneBusAway GTFS within OTP's internal model by creating new classes replacing the external classes (#2494)
 - Allow itineraries in response to be sorted by duration (#2593)
+- Add support for GTFS-flex services: flag stops, deviated-route service, and call-and-ride (#2603)
+- Fix reverse optimization bug (#2653, #2411)
 
 ## 1.3 (2018-08-03)